+++ rn-chat-view/PaginationDemoViewController.swift
@@ -202,7 +202,7 @@ final class PaginationDemoViewController: UIViewController, ChatViewControllerDe
     func chatDidEditMessage(text: String, messageId: String) {}
     func chatDidCancelInputAction(type: String) {}
     func chatDidTapAttachment() {}
-    func chatDidCompleteVoiceRecording(fileURL: URL, duration: TimeInterval) {}
+    func chatDidCompleteVoiceRecording(fileURL: URL, duration: TimeInterval, waveform: [Float]) {}
 }
 
 // MARK: - SwiftUI Wrapper